preprocessing/requirements.txt
pandas
numpy
nltk
autocorrect
unidecode
google-api-python-client 
google-auth-httplib2 
google-auth-oauthlib
matplotlib